xxxx

+++ QuanLyVanEp.WebApi/Controllers/DataService/ProductCategoryController.cs
@@ -48,40 +48,66 @@ namespace BaseApiWithIdentity.Controllers
         public async Task<ActionResult<ResponseModel>> Post(ProductCatergoryRequest value)
         {
             ProductCategoryEntity newItem = new ProductCategoryEntity();
-            newItem.Name = value.Name;
+            var found = await _context.ProductCategories.AnyAsync(c => c.Name == value.Name && c.IsDeleted == false);
+            if (!found)
+            {
+                newItem.Name = value.Name;
+                _context.ProductCategories.Add(newItem);
 
-            _context.ProductCategories.Add(newItem);
-            await _context.SaveChangesAsync();
-            return rspns.Succeed();
+                await _context.SaveChangesAsync();
+                rspns.Succeed();
+            }
+            else
+            {
+                rspns.Failed("Tên nhóm sản phẩm đã tồn tại");
+            }
+            return rspns;
         }
 
         // PUT: api/Products/5
         [HttpPut("{id}")]
         public async Task<ResponseModel> Put(int id, ProductCatergoryRequest value)
         {
-            var entity = _context.ProductCategories.Find(id);
-            if (entity == null) return rspns.NotFound();
-            entity.Name = value.Name;
+            var found = await _context.ProductCategories.AnyAsync(c => c.Name == value.Name && c.IsDeleted == false);
+            if (!found)
+            {
+                var entity = _context.ProductCategories.Find(id);
+                if (entity == null) return rspns.NotFound();
+                entity.Name = value.Name;
 
-            await _context.SaveChangesAsync();
-            return rspns.Succeed();
+                await _context.SaveChangesAsync();
+                rspns.Succeed();
+            }
+            else
+            {
+                rspns.Failed("Tên nhóm sản phẩm đã tồn tại");
+            }
+            return rspns;
         }
 
         // DELETE: api/Products/5
         [HttpDelete("{id}")]
         public async Task<ResponseModel> Delete(int id)
         {
-            var entity = await _context.ProductCategories.FindAsync(id);
-            if (entity == null)
+            var found = await _context.ProductCategories.AnyAsync(c => c.Products.Where(p => p.IsDeleted == false && p.CategoryId == id).Count() > 0);
+            if (found)
             {
-                rspns.NotFound();
-                return rspns;
+                rspns.Failed("Nhóm sản phẩm đang có sản phẩm");
             }
+            else
+            {
+                var entity = await _context.ProductCategories.FindAsync(id);
+                if (entity == null)
+                {
+                    rspns.NotFound();
+                    return rspns;
+                }
 
-            _context.ProductCategories.Remove(entity);
-            _context.SaveChanges();
+                _context.ProductCategories.Remove(entity);
+                _context.SaveChanges();
 
-            rspns.Succeed();
+                rspns.Succeed();
+            }
             return rspns;
         }
     }
